refactor: utils/constants 파일 정리
